create "lib" to start the dan node from code
